pyproject.toml
[project]
name = "snp-ai-agent-platform"
version = "0.1.0"
description = "Internal AI Agent Platform SDK and runtime scaffold."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "PyYAML>=6.0.2",
    "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "types-PyYAML>=6.0.12",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config"
testpaths = [
    "apps/runtime-api/tests",
    "packages/snp_agent_core/tests",
    "tests",
]
pythonpath = [
    "apps/runtime-api/src",
    "apps/eval-runner/src",
    "packages/snp_agent_core/src",
    "packages/snp_agent_tools/src",
    "packages/snp_agent_rag/src",
    "packages/snp_agent_memory/src",
    "packages/snp_agent_safety/src",
    "packages/snp_agent_observability/src",
    "packages/snp_agent_testing/src",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = [
    "apps/runtime-api/src",
    "apps/eval-runner/src",
    "packages/snp_agent_core/src",
    "packages/snp_agent_tools/src",
    "packages/snp_agent_rag/src",
    "packages/snp_agent_memory/src",
    "packages/snp_agent_safety/src",
    "packages/snp_agent_observability/src",
    "packages/snp_agent_testing/src",
    "tests",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "RUF",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
files = [
    "apps/runtime-api/src",
    "apps/runtime-api/tests",
    "packages/snp_agent_core/src",
    "packages/snp_agent_core/tests",
    "tests",
]
mypy_path = [
    "apps/runtime-api/src",
    "apps/eval-runner/src",
    "packages/snp_agent_core/src",
    "packages/snp_agent_tools/src",
    "packages/snp_agent_rag/src",
    "packages/snp_agent_memory/src",
    "packages/snp_agent_safety/src",
    "packages/snp_agent_observability/src",
    "packages/snp_agent_testing/src",
]
